The previous benchmark implementation was focused only on the called operations itself ignoring the part of deferred calculation that actually should be done to be able to use the calculated result.

Why is it important? To make the comparison more fair since when using Vec and its methods we get the entity that could be used immediately for acquiring the specific values or iterating over. But for `Chunk` we need to call `.as_vec` method before we can use similar things.

It changes the results significantly and could represent downsides of using `Chunk` more clearly.

Here are the results on my machine:

```
append/Chunk            time:   [1.0212 ms 1.0430 ms 1.0703 ms]
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
append/Vec              time:   [481.98 µs 483.49 µs 485.36 µs]
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

Benchmarking prepend/Chunk: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.0s, enable flat sampling, or reduce sample count to 50.
prepend/Chunk           time:   [1.7710 ms 1.7810 ms 1.7927 ms]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
prepend/Vec             time:   [16.131 ms 16.282 ms 16.461 ms]
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

Benchmarking concat/Chunk: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.1s, or reduce sample count to 90.
concat/Chunk            time:   [51.900 ms 52.134 ms 52.406 ms]
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
Benchmarking concat/Vec: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.5s, or reduce sample count to 90.
concat/Vec              time:   [50.810 ms 51.041 ms 51.352 ms]
Found 10 outliers among 100 measurements (10.00%)
  9 (9.00%) high mild
  1 (1.00%) high severe

clone/Chunk             time:   [1.8813 µs 1.9128 µs 1.9514 µs]
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) high mild
  10 (10.00%) high severe
clone/Vec               time:   [1.3566 µs 1.3694 µs 1.3856 µs]
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe

from_iter/Chunk         time:   [3.2536 µs 3.2681 µs 3.2904 µs]
Found 10 outliers among 100 measurements (10.00%)
  6 (6.00%) high mild
  4 (4.00%) high severe
from_iter/Vec           time:   [1.3630 µs 1.3896 µs 1.4259 µs]
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) high mild
  10 (10.00%) high severe
```