@tusharmath 2024-11-26
The benchmarks showcase that for almost all cases chunks are slower apart from prepend. Which contradicts the whole purpose of this data structure when in reality we are combating the performance penalty we need to deal with while working with vecs.

You won't always materialize the value as a vector. 

1. For eg: I might just need to count the total number of values or sum of value or max value etc. There are plenty of operations where materializing the chunk as vec can be skipped. 
2. There are still some cases where materialization needs to be performed only once instead of for each iteration (Like shown in the current benchmark). For eg: Concatening a large number of chunks and then materializing it once. 

Overall I don't think the premise that Chunks have to converted to Vec to do anything meaningful