@meskill 2024-11-26
> The premise for your BMs is that a chunk needs to be converted to a vector in every case. So the benchmarks showcase that for almost all instances chunks are slower apart from prepend. This contradicts the whole purpose of having this data structure when in reality we are combating the performance penalty we need to deal with while working with vecs.
> 
> In the real world, you won't always materialize the value as a vector.
> 
> 1. For eg: I might just need to count the total number of values or sum of value or max value etc. There are plenty of operations where materializing the chunk as vec can be skipped.
> 2. There are still some cases where materialization needs to be performed only once instead of for each iteration (Like shown in the current benchmark). For eg: Concatening a large number of chunks and then materializing it once.
> 3. There are also some cases where instead of materializing a chunk to a Vec, a LinkedList or a Tree or a Triė is more appropriate.
> 
> Overall, I don't think the premise that Chunks have to convert to Vec to do anything meaningful is inaccurate.

That's the point of this pr - showcase the example close to the real use case, otherwise the benchmarks could be treated as manually crafted to showcase the benefits that are unlikely achievable. 

Let's break the examples you've provided:
1. If you don't need elements why do you need any data structure at all? And even if you do use data structure to do this, how the `Chunk` could be used to do this? It seems there is no methods and the iteration over data structure will require CPU time spend anyway that is missed here, right? If you still think it's applicable in this case, let's add more benchmarks and compare it to Vec's speed
2. That's how the benchmark already implemented for append, prepend, concat - execute the operation N times and then convert it once with `as_vec`. Please, check the code if you think it's implemented wrong
3. Converting to different structures doesn't mean that would be fast because the operations are fast and we still should consider the conversion stage, right? What if creating LinkedList/Tree etc. from the start will be faster than using `Chunk`?

The current benchmark feels more like comparing iterator adapters e.g. `filter`, `map` without any consuming iterators and saying that they are faster then for loop. But we know those operations don't run actual calculation and this will be considered unfair comparison.