@tusharmath 2024-11-28
Concat comparison currently is fair. Typically in any immutable DS, concat guarantees that if two instances are concatenated, the internal order of values of the resulting DS follows the order of concatenation. It doesn't expect that each value should be accessible by index and considering its immutable, it needs to ensure that none of the two values are modified. The proposed benchmarks don't perform concatenation technically in the vector implementation. It modifies the base value with the new value which is not a fair comparison because it's causing a mutation.


The same is true for append. Appending a value only needs to guarantee that while iterating it should be the last value produced. Again the vector implementation isn't immutable and hence not an apples-to-apples comparison.